Configure AuditLogger outside AuditLogGenerationFilter
